requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
numpy==2.1.3
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
sentence-transformers==3.4.1
transformers==4.48.3
einops
timm
pillow
psycopg2-binary